Tests/LibWeb/Text/expected/DOM/range-bounding-client-rect-with-nested-text.txt

  Hello   Ladybird   Ladybird again   World    6
4
